Lint: resolve executable paths prior to execution
We were previously not resolving the binary path prior to executing it.
This was fine on Linux, but on Windows resolving the binary may add
extensions like .cmd or .exe. If the requested binary was './bin/wibble'
and there existed both a 'wibble' and 'wibble.exe' in the bin directory,
resolveBinary() returns 'wibble.exe', and it is this that needs to be
executed.